requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface-hub
python-dotenv
streamlit
